services/mock_api/requirements.txt
grpcio==1.80.0
grpcio-health-checking==1.80.0
grpcio-reflection==1.80.0
grpcio-tools==1.80.0
protobuf==6.33.6
